# Swagger documentation.json
{
   "swagger":"2.0",
   "info":{
      "version":"v1",
      "title":"DragonetWorksheetAPI"
   },
   "host":"localhost:2403",
   "schemes":[
      "http"
   ],
   "paths":{
...
      "/api/v1/bookDetail/ageingAndFeeds":{
         "post":{
            "tags":[
               "bookDetail"
            ],
            "operationId":"bookDetail_getAgeingAndFeeds",
            "consumes":[
               "application/json",
               "text/json",
               "application/x-www-form-urlencoded"
            ],
            "produces":[
               "application/json",
               "text/json"
            ],
            "parameters":[
               {
                  "name":"bookDetail",
                  "in":"body",
                  "required":true,
                  "schema":{
                     "$ref":"#/definitions/BookDetail"
                  }
               }
            ],
            "responses":{
               "200":{
                  "description":"OK",
                  "schema":{
                     "type":"object",
                     "additionalProperties":{
                        "type":"object"
                     }
                  }
               }
            }
         }
      },
...
   },
   "definitions":{
      "BookDetail":{
         "type":"object",
         "properties":{
            "centre":{
               "type":"string"
            },
            "hugo_code":{
               "type":"string"
            },
            "description":{
               "type":"string"
            },
            "ccy":{
               "type":"string"
            },
            "posn_diff":{
               "format":"int32",
               "type":"integer"
            },
            "price_diff":{
               "format":"int32",
               "type":"integer"
            },
            "pandl_diff":{
               "format":"int32",
               "type":"integer"
            },
            "pv_diff":{
               "format":"int32",
               "type":"integer"
            },
            "posn_age_break":{
               "format":"int32",
               "type":"integer"
            },
            "narrative":{
               "type":"string"
            },
            "last_upd_username":{
               "type":"string"
            },
            "last_upd_date":{
               "type":"string"
            },
            "rec_source":{
               "type":"string"
            },
            "dt_loaded":{
               "type":"string"
            },
            "bo_source":{
               "type":"string"
            },
            "fo_source":{
               "type":"string"
            },
            "trading_book":{
               "type":"string"
            },
            "aries_book":{
               "type":"string"
            },
            "book_desc":{
               "type":"string"
            },
            "prod_code":{
               "type":"string"
            },
            "isin":{
               "type":"string"
            },
            "multi":{
               "format":"double",
               "type":"number"
            },
            "fctr":{
               "format":"double",
               "type":"number"
            },
            "bo_posn":{
               "format":"double",
               "type":"number"
            },
            "bo_price":{
               "format":"double",
               "type":"number"
            },
            "bo_value":{
               "format":"double",
               "type":"number"
            },
            "bo_pandl":{
               "format":"double",
               "type":"number"
            },
            "fo_posn":{
               "format":"double",
               "type":"number"
            },
            "fo_price":{
               "format":"double",
               "type":"number"
            },
            "fo_value":{
               "format":"double",
               "type":"number"
            },
            "fo_pandl":{
               "format":"double",
               "type":"number"
            },
            "prod_type":{
               "type":"string"
            },
            "prod_type_desc":{
               "type":"string"
            },
            "ticker":{
               "format":"double",
               "type":"number"
            },
            "pool_factor":{
               "format":"double",
               "type":"number"
            },
            "price_age_break":{
               "format":"int32",
               "type":"integer"
            },
            "pandl_age_break":{
               "format":"int32",
               "type":"integer"
            },
            "gmi_trade_price_multi":{
               "format":"double",
               "type":"number"
            },
            "gmi_strike_price_multi":{
               "format":"double",
               "type":"number"
            },
            "bo_dividend_pandl":{
               "format":"double",
               "type":"number"
            },
            "ticker_code":{
               "type":"string"
            },
            "product_category":{
               "type":"string"
            }
         }
      },
...
   }
}